# api/swagger/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "version": "2.0.0",
        "title": "Quick Dashboard API",
        "description": "Manages data for the Dashboard",
        "contact": {
            "name": "Benjamin Weare"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "https://localhost:8000",
            "description": "Local Development"
        },
        {
            "url": "https://intranet-api.local",
            "description": "Hosted Internally"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "APIKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-KEY"
            }
        },
        "schemas": {
            "id": {
                "type": "string"
            },
...
            "weatherOption": {
                "type": "string",
                "enum": ["daily"]
            },
...
            "error": {
                "required": ["message"],
                "properties": {
                    "message": { "type": "string" }
                }
            },
...
            "sticky": {
                "properties": {
                    "message": { "type": "string" },
                    "itemsLength": { "type": "integer" },
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "_id": { "type": "object" },
                                "dateTime": { "type": "string" },
                                "top": { "type": "number" },
                                "left": { "type": "number" },
                                "title": { "type": "string" },
                                "content": { "type": "string" },
                                "color": { "type": "string" },
                                "showColor": { "type": "string" }
                            }
                        }
                    }
                }
            },
            "stickyId": {
                "properties": {
                    "message": { "type": "string" },
                    "response": {
                        "type": "object",
                        "properties": {
                            "_id": { "type": "string" },
                            "dateTime": { "type": "string" },
                            "top": { "type": "integer" },
                            "left": { "type": "integer" },
                            "title": { "type": "string" },
                            "content": { "type": "string" },
                            "color": { "type": "string" },
                            "showColor": { "type": "string" }
                        }
                    }
                }
            },
            "sync": { "properties": { "message": { "type": "string" } } },
...
            "stickyBody": {
                "type": "object",
                "properties": {
                    "author": { "type": "string" }
                }
            },
            "stickyEdit": {
                "type": "object",
                "required": ["top", "left", "title", "content", "color"],
                "properties": {
                    "top": { "type": "number" },
                    "left": { "type": "number" },
                    "title": { "type": "string" },
                    "content": { "type": "string" },
                    "color": { "type": "string" }
                }
            },
...
            "weather": {
                "type": "object",
                "properties": {
                    "message": { "type": "string" },
                    "itemsLength": { "type": "integer" },
                    "items": { "type": "object" }
                }
            },
...
        }
    },
    "security": [{ "APIKeyAuth": [] }],
    "paths": {
        "/sticky": {
            "get": {
                "description": "Returns all notes",
                "tags": ["Sticky Notes"],
                "summary": "Returns all notes",
                "x-eov-operation-handler": "sticky",
                "operationId": "getStickies",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sticky"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/stickyBody"
                            }
                        }
                    },
                    "required": true
                },
                "tags": ["Sticky Notes"],
                "summary": "Returns a new note",
                "x-eov-operation-handler": "sticky",
                "operationId": "postSticky",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/stickyId"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sticky/{id}": {
            "get": {
                "description": "Returns a note",
                "tags": ["Sticky Notes"],
                "summary": "Returns a note",
                "x-eov-operation-handler": "sticky",
                "operationId": "getSticky",
                "parameters": [
                    {
                        "schema": { "$ref": "#/components/schemas/id" },
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/stickyId"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Edits a note",
                "tags": ["Sticky Notes"],
                "summary": "Edits a note",
                "x-eov-operation-handler": "sticky",
                "operationId": "patchSticky",
                "parameters": [
                    {
                        "schema": { "$ref": "#/components/schemas/id" },
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/stickyEdit"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sticky"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Edits a note",
                "tags": ["Sticky Notes"],
                "summary": "Edits a note",
                "x-eov-operation-handler": "sticky",
                "operationId": "deleteSticky",
                "parameters": [
                    {
                        "schema": { "$ref": "#/components/schemas/id" },
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sticky"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/news": {
            "get": {
                "tags": ["News"],
                "summary": "Returns all stored news",
                "x-eov-operation-handler": "news",
                "operationId": "getNews",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": { "type": "string" },
                                        "itemsLength": { "type": "integer" },
                                        "items": { "type": "array" }
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": ["message"],
                                    "properties": {
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/weather": {
            "get": {
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/components/schemas/weatherOption"
                        },
                        "in": "query",
                        "name": "date",
                        "required": true
                    }
                ],
                "tags": ["Weather"],
                "summary": "Returns all stored weather",
                "x-eov-operation-handler": "weather",
                "operationId": "getWeather",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/weather"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sync": {
            "get": {
                "tags": ["Sync"],
                "summary": "Syncs news and weather from API endpoints",
                "x-eov-operation-handler": "sync",
                "operationId": "getSync",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/sync"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}